optimize imports, update some langs, cleanup, tweaked textures

+++ src/main/java/teamport/aether/lookup/LookupFuelFreezer.java
@@ -38,7 +38,7 @@ public class LookupFuelFreezer {
     }
 
     public int getFuelYield(int id) {
-        return this.fuelList.get(id) == null ? 0 : (Integer)this.fuelList.get(id);
+        return this.fuelList.get(id) == null ? 0 : this.fuelList.get(id);
     }
 
     public Map<Integer, Integer> getFuelList() {
